Replace "project" in the setup.xml with "group" and "module".
This commit requires projects to change the setup.xml file.

The "muhkuh_buildsystem" node has no longer a "project" attribute. It is
replaced by the "group" and "module" attributes.

Old:
  <muhkuh_buildsystem project="org.muhkuh.lua-uart">

New:
  <muhkuh_buildsystem group="org.muhkuh.lua" module="uart">

The values of the "group" and "module" attributes are available in the
SConstruct files as "PROJECT_GROUP" and "PROJECT_MODULE".